+++ elkm1/lights.py
@@ -2,7 +2,8 @@
 
 from .const import Max, TextDescriptions
 from .elements import Element, Elements
-from .message import add_message_handler, ps_encode
+from .message import add_message_handler, ps_encode, pc_encode, pf_encode,
+                     pn_encode, pt_encode
 
 
 class Light(Element):