+++ elkm1/tasks.py
@@ -1,4 +1,6 @@
 """Definition of an ElkM1 Task"""
+from time import time
+
 from .const import Max, TextDescriptions
 from .elements import Element, Elements
 from .message import tn_encode
@@ -9,6 +11,7 @@ class Task(Element):
 
     def __init__(self, index): # pylint: disable=useless-super-delegation
         super().__init__(index)
+        self.last_change = None
 
     def activate(self):
         """(Helper) Activate task"""
@@ -18,7 +21,11 @@ class Tasks(Elements):
     """Handling for multiple tasks"""
     def __init__(self, elk):
         super().__init__(elk, Task, Max.TASKS.value)
+        add_message_handler('TC', self._tc_handler)
 
     def sync(self):
         """Retrieve tasks from ElkM1"""
         self.get_descriptions(TextDescriptions.TASK.value)
+
+    def _tc_handler(self, task):
+        self.elements[task].setattr('last_change', time())