+++ elkm1/settings.py
@@ -11,6 +11,10 @@ class Setting(Element):
         self.value_format = None
         self.value = None
 
+    def set(self, value_format, value):
+        """(Helper) Set custom value."""
+        self.elk.send(cw_encode(self._index, value, value_format))
+
 class Settings(Elements):
     """Handling for multiple custom values"""
     def __init__(self, elk):