+++ bin/cmdr.py
@@ -9,9 +9,20 @@ import inspect
 import re
 import threading
 import urwid
+from importlib import import_module
 
 import elkm1.message
 
+
+def parse_subcommand(self, line):
+    converted = []
+    for arg in args:
+        try:
+            i = int(arg)
+            converted.append(i)
+        except ValueError:
+            converted.append(arg)
+    return converted
 
 def parse_range(rng, max):
     # e.g. rng = "<3, 45, 46, 48-51, 77"
@@ -30,116 +41,124 @@ def parse_range(rng, max):
             raise ValueError('Unknown range type: "%s"'%x)
     return ids
 
-class UnknownCommand(Exception):
-    def __init__(self,cmd):
-        Exception.__init__(self,'Uknown command: %s'%cmd)
+def parse_element_command(line, max):
+    match = re.match( '([\d,\- <*]+)(\w*.*)', line)
+    ids = parse_range(match.groups[0])
 
-class Command(object):
-    """Base class to manage commands."""
+def find_class(module, class_name):
+    for class_ in inspect.getmembers(module, inspect.isclass):
+        if class_[0] == class_name:
+            return class_
+    return None
 
-    def __init__(self):
+def get_helpers(element, clas):
+    helpers = {}
+    class_ = find_class(import_module('elkm1.' + element), clas)
+    if class_:
+        for function_name, fn in inspect.getmembers(class_[1], inspect.isfunction):
+            if fn.__doc__ and fn.__doc__.startswith('(Helper)'):
+                params = [p for p in inspect.signature(fn).parameters]
+                if params[0] == 'self':
+                    del params[0]
+                params = ' '.join(['<'+p+'>' for p in params])
+                helpers[function_name] = (fn,
+                        '{} {}'.format(function_name,params),
+                        fn.__doc__[8:])
+    return helpers
+
+class Commands():
+    def __init__(self, elk):
+        self.elk = elk
+
         self._quit_cmd = ['q', 'quit', 'exit']
         self._help_cmd = ['?', 'help', 'h']
 
-        self.encoders = {}
+        self.msg_cmds = {}
         for fn_name, fn in inspect.getmembers(elkm1.message, inspect.isfunction):
             if not fn_name.endswith('_encode'):
                 continue
-            self.encoders[fn_name[0:2]] = fn
+            cmd = fn_name[0:2]
+            params = [p for p in inspect.signature(fn).parameters]
+            params = ' '.join(['<'+p+'>' for p in params])
+            self.msg_cmds[cmd] = (fn, '{} {}'.format(cmd, params), fn.__doc__)
 
-    def __call__(self,line):
-        tokens=line.split()
-        cmd=tokens[0].lower()
-        args=tokens[1:]
+        self.element_cmds = {}
+        self.subcommands = {}
+        for element in elk.element_list:
+            if element == 'panel':
+                fn = self.panel_print
+                cmd = element
+            else:
+                fn = self.print_elements
+                cmd = element[:-1]
+            self.element_cmds[cmd] = (fn,
+                             '{} <range list> [subcommand]'.format(cmd),
+                             'Displays internal state of {}'.format(element),
+                              get_helpers(element,cmd.capitalize()))
+
+    def __call__(self, line):
+        tokens = line.split()
+        cmd = tokens[0].lower()
+        args = tokens[1:]
 
         if cmd in self._quit_cmd:
             return Commander.Exit
 
-        print("#blue#{:s} {:s}".format(cmd, ' '.join(args)))
+        print("#blue#{}".format(line))
 
         if cmd in self._help_cmd:
-            return self.help(args[0] if args else None)
+            return self.help(cmd, args)
 
-        if hasattr(self, 'do_'+cmd):
-            return getattr(self, 'do_'+cmd)(*args)
+        if cmd in self.msg_cmds:
+            return self.encoder(cmd, args)
 
-        if cmd in self.encoders:
-            converted = []
-            for arg in args:
-                try:
-                    i = int(arg)
-                    converted.append(i)
-                except ValueError:
-                    converted.append(arg)
-            self.encoder(self.encoders[cmd], *converted)
+        if cmd in self.element_cmds:
+            return self.element_cmds[cmd][0](cmd, args)
+
+        return '#error#Unknown command: {}'.format(cmd)
+
+    def help(self, cmd, args):
+        if len(args) == 0:
+            res = '#green#Type "[?|h|help] <command>" to get more help\n'
+            res += 'Type "[q|quit|exit]" to quit program\n'
+            res += 'Element display commands:\n  {}\n\n'.format(
+                ' '.join(list(self.element_cmds.keys())))
+            cl = [fnname for fnname in self.msg_cmds]
+            res += 'Send message commands:\n  {}\n'.format(
+                ' '.join(sorted(cl)))
         else:
-            raise UnknownCommand(cmd)
-
-    def help(self,cmd=None):
-        def std_help():
-            qc='|'.join(self._quit_cmd)
-            hc ='|'.join(self._help_cmd)
-
-            res='#green#Type [%s] <command> to get more help about particular command\n' % hc
-            res+='Type [%s] to quit program\n' % qc
-
-            cl = [name[3:] for name in dir(self) if name.startswith('do_')]
-            cl += [fnname for fnname in self.encoders]
-            res += 'Available commands: %s' %(' '.join(sorted(cl)))
-            return res
-
-        if not cmd:
-            return std_help()
-        else:
-            if cmd in self.encoders:
-                params = [p for p in inspect.signature(self.encoders[cmd]).parameters]
-                doc = "#green#{:s} {:s}\n".format(cmd, " ".join(params))
-                doc += self.encoders[cmd].__doc__
-                return doc
+            help_for = args[0]
+            if help_for in self.msg_cmds:
+                res = "#green#{}\n{}".format(self.msg_cmds[help_for][1], 
+                                        self.msg_cmds[help_for][2])
+            elif help_for in self.element_cmds:
+                res = "#green#{}\n{}".format(self.element_cmds[help_for][1], 
+                                        self.element_cmds[help_for][2])
+                for k, v in self.element_cmds[help_for][3].items():
+                    res += '\nSubcommand: {}\n{}'.format(v[1], v[2])
             else:
-                try:
-                    fn=getattr(self,'do_'+cmd)
-                    doc=fn.__doc__
-                    return doc or '#red#No documentation available for %s'%cmd
-                except AttributeError:
-                    return std_help()
- 
-class ElkCommands(Command):
-    def set_elk(self, elk):
-        self.elk = elk
+                res = "#error#Unknown command: {}".format(help_for)
+        return res
 
-    def do_raw(self, *args):
-        """raw - send bytes typed to Elk"""
-        self.elk.send(elkm1.message.MessageEncode(args[0], None))
+    def print_elements(self, cmd, args):
+        element_list = getattr(self.elk, cmd+'s', None)
+        if not element_list:
+            element_list = getattr(self.elk, cmd)
+        for i in parse_range(" ".join(args), element_list.max_elements):
+            print(element_list[i])
 
-    def display_objects(self, object_list, *args):
-        for i in parse_range(" ".join(args), object_list.max_elements):
-            print(object_list[i])
-
-    def do_zone(self, *args):
-        self.display_objects(self.elk.zones, *args)
-    def do_light(self, *args):
-        self.display_objects(self.elk.lights, *args)
-    def do_area(self, *args):
-        self.display_objects(self.elk.areas, *args)
-    def do_keypad(self, *args):
-        self.display_objects(self.elk.keypads, *args)
-    def do_user(self, *args):
-        self.display_objects(self.elk.users, *args)
-    def do_task(self, *args):
-        self.display_objects(self.elk.tasks, *args)
-    def do_output(self, *args):
-        self.display_objects(self.elk.outputs, *args)
-    def do_setting(self, *args):
-        self.display_objects(self.elk.settings, *args)
-    def do_counter(self, *args):
-        self.display_objects(self.elk.counters, *args)
-    def do_thermostat(self, *args):
-        self.display_objects(self.elk.thermostats, *args)
-    def do_panel(self, *args):
+    def panel_print(self, cmd, args):
         print(self.elk.panel)
-    def encoder(self, fn, *args):
-        self.elk.send(fn(*args))
+
+    def encoder(self, cmd, args):
+        converted = []
+        for arg in args:
+            try:
+                i = int(arg)
+                converted.append(i)
+            except ValueError:
+                converted.append(arg)
+        self.elk.send(self.msg_cmds[cmd][0](*converted))
 
 class FocusMixin(object):
     def mouse_event(self, size, event, button, x, y, focus):